+++ packages/core/src/domain/report/reportObservable.spec.ts
@@ -21,7 +21,6 @@ describe('report observable', () => {
   })
 
   afterEach(() => {
-    reportingObserver.reset()
     consoleSubscription.unsubscribe()
   })
   ;[RawReportType.deprecation, RawReportType.intervention].forEach((type) => {