+++ packages/core/src/tools/experimentalFeatures.ts
@@ -18,7 +18,6 @@ export enum ExperimentalFeature {
   REMOTE_CONFIGURATION = 'remote_configuration',
   UPDATE_VIEW_NAME = 'update_view_name',
   LONG_ANIMATION_FRAME = 'long_animation_frame',
-  VIEW_SPECIFIC_CONTEXT = 'view_specific_context',
   ANONYMOUS_USER_TRACKING = 'anonymous_user_tracking',
 }